+++ .gitignore
@@ -42,4 +42,3 @@ bin/
 .DS_Store
 
 application.yml
-post.http
