+++ src/main/java/org/sopt/dto/PostDetailResponse.java
@@ -1,4 +1,4 @@
-package org.sopt.dto.res;
+package org.sopt.dto;
 
 public record PostDetailResponse(Long id, String title) {
     public static PostDetailResponse of(final Long id, final String title) {
