+++ src/main/java/org/sopt/dto/PostListResponse.java
@@ -1,0 +1,11 @@
+package org.sopt.dto;
+
+import org.sopt.domain.Post;
+
+import java.util.List;
+
+public record PostListResponse(List<Post> posts) {
+    public static PostListResponse of(final List<Post> posts) {
+        return new PostListResponse(posts);
+    }
+}
