+++ src/main/java/org/sopt/service/PostService.java
@@ -1,9 +1,10 @@
 package org.sopt.service;
 
 import org.sopt.domain.Post;
-import org.sopt.dto.req.PostCreateRequest;
-import org.sopt.dto.res.PostCreateResponse;
-import org.sopt.dto.res.PostDetailResponse;
+import org.sopt.dto.PostCreateRequest;
+import org.sopt.dto.PostCreateResponse;
+import org.sopt.dto.PostDetailResponse;
+import org.sopt.dto.PostListResponse;
 import org.sopt.repository.PostRepository;
 import org.springframework.stereotype.Service;
 
@@ -24,9 +25,9 @@ public class PostService {
     }
 
 
-//    public List<Post> getAllPosts() {
-//        return postRepository.findAll();
-//    }
+    public PostListResponse getAllPosts() {
+        return PostListResponse.of(postRepository.findAll());
+    }
 
     public PostDetailResponse getPost(final Long id) {
         Post post = postRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("해당 게시글이 없습니다. id=" + id));
